built-in/nlp/bert_base_finetune_msra_ner/models/requirements.txt
tqdm
boto3
pytorch_pretrained_bert==0.4.0